Updating other components to use PollStore
