+++ src/App.svelte
@@ -13,22 +13,7 @@
 		activeItem = e.detail;
 	};
 
-	const handleVote = (e) => {
-		const { id, option } = e.detail;
-		let copiedPolls = [...polls];
-		let upvotedPoll = copiedPolls.find((poll) => poll.id === id);
-		if (option === "a") {
-			upvotedPoll.votesA++;
-		} else if (option === "b") {
-			upvotedPoll.votesB++;
-		}
-
-		polls = copiedPolls;
-	};
-
 	const handleAddPoll = (e) => {
-		const poll = e.detail;
-		polls = [poll, ...polls];
 		activeItem = "Current Polls";
 	};
 </script>
@@ -37,7 +22,7 @@
 <main>
 	<Tabs {activeItem} {items} on:tabChange={tabChange} />
 	{#if activeItem === "Current Polls"}
-		<PollList on:vote={handleVote} />
+		<PollList />
 	{:else if activeItem === "Add New Poll"}
 		<CreatePollForm on:addPoll={handleAddPoll} />
 	{/if}
