+++ src/components/PollList.svelte
@@ -6,7 +6,7 @@
 <div class="poll-list">
     {#each $PollStore as poll (poll.id)}
         <div>
-            <PollDetails {poll} on:vote />
+            <PollDetails {poll} />
         </div>
     {/each}
 </div>
